# apps/server/swagger-spec.yaml
openapi: 3.0.3
info:
  title: API doc
  description: |-
    This is Test purpose
  termsOfService: http://swagger.io/terms/
  contact:
    email: aswin@cartrabbit.com
  version: 1.0.11
  # ----------------------------------------------------------------- tags ----------------------------------------------------------------------------
...
paths:
...
  /users:
    post:
      summary: Create a user
      operationId: createUser
      tags:
        - Users
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  example: "John Doe"
                email:
                  type: string
                  example: "johndoe@example.com"
                role:
                  type: string
                  example: "admin"
                authenticationMethod:
                  type: string
                  enum:
                    - GOOGLE
                    - MAGIC_LINK
              required:
                - name
                - email
                - role
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Invalid request body
        '500':
          description: Internal server error
  /users/{userId}:
    parameters:
      - in: path
        name: userId
        required: true
        schema:
          type: string
        description: ID of the user to retrieve
    get:
      summary: Retrieve a user by ID
      operationId: getUserById
      tags:
        - Users
      responses:
        '200':
          description: User found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '404':
          description: User not found
        '500':
          description: Internal server error

    put:
      summary: Update a user
      operationId: updateUser
      tags:
        - Users
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
      responses:
        '200':
          description: User updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Invalid request body
        '404':
          description: User not found
        '500':
          description: Internal server error

    delete:
      summary: Delete a user
      operationId: deleteUser
      tags:
        - Users
      responses:
        '204':
          description: User deleted successfully
        '404':
          description: User not found
        '500':
          description: Internal server error

  # OrgMemberInvite
...
components:
  schemas:
    User:
      type: object
      properties:
        id:
          type: string
          example: "ck2j49r2a0d1m0741gzhjo41v"
        email:
          type: string
          example: "user@example.com"
        authenticationMethod:
          type: string
          enum:
            - GOOGLE
            - MAGIC_LINK
        authenticationProfile:
          type: object
        name:
          type: string
          example: "John Doe"
        avatarUrl:
          type: string
          format: uri
          example: "http://example.com/avatar.jpg"
        admin:
          type: boolean
          default: false
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        deletedAt:
          type: string
          format: date-time
        OrgMemberInvite:
          type: array
          items:
            $ref: '#/components/schemas/OrgMemberInvite'
        OrgMember:
          type: array
          items:
            $ref: '#/components/schemas/OrgMember'
    OrgMemberInvite:
      type: object
      properties:
        id:
          type: string
          example: "ck2j49r2a0d1m0741gzhjo41v"
        token:
          type: string
          example: "ck2j49r2a0d1m0741gzhjo41v"
        email:
          type: string
          example: "invitee@example.com"
        role:
          type: string
          enum:
            - ADMIN
            - MEMBER
          example: "MEMBER"
        organizationId:
          type: string
          example: "ck2j49r2a0d1m0741gzhjo41v"
        inviterId:
          type: string
          example: "ck2j49r2a0d1m0741gzhjo41v"
        createdAt:
          type: string
          format: date-time
          example: "2023-07-18T12:34:56Z"
        updatedAt:
          type: string
          format: date-time
          example: "2023-07-18T12:34:56Z"
...
    OrgMember:
      type: object
      properties:
        id:
          type: string
          example: "ck2j49r2a0d1m0741gzhjo41v"
        organizationId:
          type: string
          example: "ck2j49r2a0d1m0741gzhjo41v"
        userId:
          type: string
          example: "ck2j49r2a0d1m0741gzhjo41v"
        role:
          type: string
          enum:
            - ADMIN
            - MEMBER
          example: "MEMBER"
        createdAt:
          type: string
          format: date-time
          example: "2023-07-18T12:34:56Z"
        updatedAt:
          type: string
          format: date-time
          example: "2023-07-18T12:34:56Z"
...
  securitySchemes:
    api_key:
      type: apiKey
      name: api_key
      in: header